+++ src/main/java/com/hotspot/livfit/exercise/controller/LungeController.java
@@ -1,4 +1,4 @@
-package com.hotspot.livfit.exercise.entity.controller;
+package com.hotspot.livfit.exercise.controller;
 
 import java.util.List;
 
@@ -8,23 +8,23 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.hotspot.livfit.exercise.entity.dto.Record;
-import com.hotspot.livfit.exercise.entity.entity.Squat;
-import com.hotspot.livfit.exercise.entity.service.ExerciseService;
+import com.hotspot.livfit.exercise.dto.Record;
+import com.hotspot.livfit.exercise.entity.Lunge;
+import com.hotspot.livfit.exercise.service.ExerciseService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 
 @RestController
-@RequestMapping("/squat")
+@RequestMapping("/lunge")
 @RequiredArgsConstructor
 @Slf4j
-public class SquatController {
+public class LungeController {
   private final ExerciseService exerciseService;
-  // 스쿼트 기록 저장 엔드포인트
+  // 런지 기록 저장 엔드포인트
   /*
-   * URL : /api/squat/record
+   * URL : /api/lunge/record
    * HTTP Method: POST
    * 요청 JSON 형식 :
    * {
@@ -37,7 +37,7 @@ public class SquatController {
    *   "good" : "5"
    * }
    * */
-  @Operation(summary = "기록저장", description = "푸쉬업 기록 저장")
+  @Operation(summary = "기록저장", description = "런지 기록 저장")
   @ApiResponses(
       value = {
         @ApiResponse(responseCode = "200", description = "기록 완료."),
@@ -47,8 +47,8 @@ public class SquatController {
   @PostMapping("/record")
   public ResponseEntity<?> saveRecord(@RequestBody Record record) {
     try {
-      Squat squat =
-          exerciseService.saveRecordSquat(
+      Lunge lunge =
+          exerciseService.saveRecordLunge(
               record.getToken(),
               record.getUsername(),
               record.getTimer_sec(),
@@ -56,21 +56,21 @@ public class SquatController {
               record.getPerfect(),
               record.getGreat(),
               record.getGood());
-      return ResponseEntity.ok(squat);
+      return ResponseEntity.ok(lunge);
 
     } catch (RuntimeException e) {
       log.error(
-          "Error during saving squat record in controller /api/squat/record: {}", e.getMessage());
+          "Error during saving Lunge record in controller /api/lunge/record: {}", e.getMessage());
       return ResponseEntity.badRequest().body(e.getMessage());
     }
   }
 
-  // 스쿼트 기록 가져오는 엔드포인트
+  // 런지 기록 가져오는 엔드포인트
   /*
-   * URL : /api/squat/all
+   * URL : /api/lunge/all
    * HTTP Method: GET
    * */
-  @Operation(summary = "기록 가져오기", description = "스쿼트 기록 조회")
+  @Operation(summary = "기록 가져오기", description = "런지 기록 조회")
   @ApiResponses(
       value = {
         @ApiResponse(responseCode = "200", description = " 가져오기 완료."),
@@ -78,13 +78,12 @@ public class SquatController {
         @ApiResponse(responseCode = "500", description = "서버 에러.")
       })
   @GetMapping("/all")
-  public ResponseEntity<List<Squat>> getAllRecords() {
+  public ResponseEntity<List<Lunge>> getAllRecords() {
     try {
-      List<Squat> squats = exerciseService.getAllSquat();
-      return ResponseEntity.ok(squats);
+      List<Lunge> lunges = exerciseService.getAllLunge();
+      return ResponseEntity.ok(lunges);
     } catch (RuntimeException e) {
-      log.error(
-          "Error during fetching pushup records in controller /squat/all: {}", e.getMessage());
+      log.error("Error during fetching Lunge records in controller /lunge/all: {}", e.getMessage());
       return ResponseEntity.status(500).body(null);
     }
   }
