+++ src/main/java/com/hotspot/livfit/exercise/controller/PushupController.java
@@ -1,4 +1,4 @@
-package com.hotspot.livfit.exercise.entity.controller;
+package com.hotspot.livfit.exercise.controller;
 
 import java.util.List;
 
@@ -8,23 +8,23 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.hotspot.livfit.exercise.entity.dto.Record;
-import com.hotspot.livfit.exercise.entity.entity.Lunge;
-import com.hotspot.livfit.exercise.entity.service.ExerciseService;
+import com.hotspot.livfit.exercise.dto.Record;
+import com.hotspot.livfit.exercise.entity.Pushup;
+import com.hotspot.livfit.exercise.service.ExerciseService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 
 @RestController
-@RequestMapping("/lunge")
+@RequestMapping("/pushup")
 @RequiredArgsConstructor
 @Slf4j
-public class LungeController {
+public class PushupController {
   private final ExerciseService exerciseService;
-  // 런지 기록 저장 엔드포인트
+  // 푸쉬업 기록 저장 엔드포인트
   /*
-   * URL : /api/lunge/record
+   * URL : /api/pushup/record
    * HTTP Method: POST
    * 요청 JSON 형식 :
    * {
@@ -37,7 +37,7 @@ public class LungeController {
    *   "good" : "5"
    * }
    * */
-  @Operation(summary = "기록저장", description = "런지 기록 저장")
+  @Operation(summary = "기록저장", description = "푸쉬업 기록 저장")
   @ApiResponses(
       value = {
         @ApiResponse(responseCode = "200", description = "기록 완료."),
@@ -47,8 +47,8 @@ public class LungeController {
   @PostMapping("/record")
   public ResponseEntity<?> saveRecord(@RequestBody Record record) {
     try {
-      Lunge lunge =
-          exerciseService.saveRecordLunge(
+      Pushup pushup =
+          exerciseService.saveRecordPushup(
               record.getToken(),
               record.getUsername(),
               record.getTimer_sec(),
@@ -56,21 +56,21 @@ public class LungeController {
               record.getPerfect(),
               record.getGreat(),
               record.getGood());
-      return ResponseEntity.ok(lunge);
+      return ResponseEntity.ok(pushup);
 
     } catch (RuntimeException e) {
       log.error(
-          "Error during saving Lunge record in controller /api/lunge/record: {}", e.getMessage());
+          "Error during saving Pushup record in controller /api/pushup/record: {}", e.getMessage());
       return ResponseEntity.badRequest().body(e.getMessage());
     }
   }
 
-  // 런지 기록 가져오는 엔드포인트
+  // 푸쉬업 기록 가져오는 엔드포인트
   /*
-   * URL : /api/lunge/all
+   * URL : /api/pushup/all
    * HTTP Method: GET
    * */
-  @Operation(summary = "기록 가져오기", description = "런지 기록 조회")
+  @Operation(summary = "기록 가져오기", description = "푸쉬업 기록 조회")
   @ApiResponses(
       value = {
         @ApiResponse(responseCode = "200", description = " 가져오기 완료."),
@@ -78,12 +78,13 @@ public class LungeController {
         @ApiResponse(responseCode = "500", description = "서버 에러.")
       })
   @GetMapping("/all")
-  public ResponseEntity<List<Lunge>> getAllRecords() {
+  public ResponseEntity<List<Pushup>> getAllRecords() {
     try {
-      List<Lunge> lunges = exerciseService.getAllLunge();
-      return ResponseEntity.ok(lunges);
+      List<Pushup> pushups = exerciseService.getAllPushup();
+      return ResponseEntity.ok(pushups);
     } catch (RuntimeException e) {
-      log.error("Error during fetching Lunge records in controller /lunge/all: {}", e.getMessage());
+      log.error(
+          "Error during fetching pushup records in controller /pushup/all: {}", e.getMessage());
       return ResponseEntity.status(500).body(null);
     }
   }
