+++ src/main/java/com/hotspot/livfit/exercise/controller/SquatController.java
@@ -1,4 +1,4 @@
-package com.hotspot.livfit.exercise.entity.controller;
+package com.hotspot.livfit.exercise.controller;
 
 import java.util.List;
 
@@ -8,23 +8,23 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.hotspot.livfit.exercise.entity.dto.Record;
-import com.hotspot.livfit.exercise.entity.entity.Pushup;
-import com.hotspot.livfit.exercise.entity.service.ExerciseService;
+import com.hotspot.livfit.exercise.dto.Record;
+import com.hotspot.livfit.exercise.entity.Squat;
+import com.hotspot.livfit.exercise.service.ExerciseService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 
 @RestController
-@RequestMapping("/pushup")
+@RequestMapping("/squat")
 @RequiredArgsConstructor
 @Slf4j
-public class PushupController {
+public class SquatController {
   private final ExerciseService exerciseService;
-  // 푸쉬업 기록 저장 엔드포인트
+  // 스쿼트 기록 저장 엔드포인트
   /*
-   * URL : /api/pushup/record
+   * URL : /api/squat/record
    * HTTP Method: POST
    * 요청 JSON 형식 :
    * {
@@ -47,8 +47,8 @@ public class PushupController {
   @PostMapping("/record")
   public ResponseEntity<?> saveRecord(@RequestBody Record record) {
     try {
-      Pushup pushup =
-          exerciseService.saveRecordPushup(
+      Squat squat =
+          exerciseService.saveRecordSquat(
               record.getToken(),
               record.getUsername(),
               record.getTimer_sec(),
@@ -56,21 +56,21 @@ public class PushupController {
               record.getPerfect(),
               record.getGreat(),
               record.getGood());
-      return ResponseEntity.ok(pushup);
+      return ResponseEntity.ok(squat);
 
     } catch (RuntimeException e) {
       log.error(
-          "Error during saving Pushup record in controller /api/pushup/record: {}", e.getMessage());
+          "Error during saving squat record in controller /api/squat/record: {}", e.getMessage());
       return ResponseEntity.badRequest().body(e.getMessage());
     }
   }
 
-  // 푸쉬업 기록 가져오는 엔드포인트
+  // 스쿼트 기록 가져오는 엔드포인트
   /*
-   * URL : /api/pushup/all
+   * URL : /api/squat/all
    * HTTP Method: GET
    * */
-  @Operation(summary = "기록 가져오기", description = "푸쉬업 기록 조회")
+  @Operation(summary = "기록 가져오기", description = "스쿼트 기록 조회")
   @ApiResponses(
       value = {
         @ApiResponse(responseCode = "200", description = " 가져오기 완료."),
@@ -78,13 +78,13 @@ public class PushupController {
         @ApiResponse(responseCode = "500", description = "서버 에러.")
       })
   @GetMapping("/all")
-  public ResponseEntity<List<Pushup>> getAllRecords() {
+  public ResponseEntity<List<Squat>> getAllRecords() {
     try {
-      List<Pushup> pushups = exerciseService.getAllPushup();
-      return ResponseEntity.ok(pushups);
+      List<Squat> squats = exerciseService.getAllSquat();
+      return ResponseEntity.ok(squats);
     } catch (RuntimeException e) {
       log.error(
-          "Error during fetching pushup records in controller /pushup/all: {}", e.getMessage());
+          "Error during fetching pushup records in controller /squat/all: {}", e.getMessage());
       return ResponseEntity.status(500).body(null);
     }
   }
