--- src/main/java/com/hotspot/livfit/exercise/entity/TestEntity.java
@@ -1,3 +1,0 @@
-package com.hotspot.livfit.exercise.entity;
-
-public class TestEntity {}
