+++ src/main/java/com/hotspot/livfit/exercise/repository/LungeRepository.java
@@ -1,7 +1,7 @@
-package com.hotspot.livfit.exercise.entity.repository;
+package com.hotspot.livfit.exercise.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.hotspot.livfit.exercise.entity.entity.Lunge;
+import com.hotspot.livfit.exercise.entity.Lunge;
 
 public interface LungeRepository extends JpaRepository<Lunge, Long> {}
