+++ src/main/java/com/hotspot/livfit/exercise/repository/PushupRepository.java
@@ -1,7 +1,7 @@
-package com.hotspot.livfit.exercise.entity.repository;
+package com.hotspot.livfit.exercise.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.hotspot.livfit.exercise.entity.entity.Pushup;
+import com.hotspot.livfit.exercise.entity.Pushup;
 
 public interface PushupRepository extends JpaRepository<Pushup, Long> {}
