+++ src/main/java/com/hotspot/livfit/exercise/repository/SquatRepository.java
@@ -1,7 +1,7 @@
-package com.hotspot.livfit.exercise.entity.repository;
+package com.hotspot.livfit.exercise.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.hotspot.livfit.exercise.entity.entity.Squat;
+import com.hotspot.livfit.exercise.entity.Squat;
 
 public interface SquatRepository extends JpaRepository<Squat, Long> {}
