+++ src/main/java/com/hotspot/livfit/exercise/service/ExerciseService.java
@@ -1,4 +1,4 @@
-package com.hotspot.livfit.exercise.entity.service;
+package com.hotspot.livfit.exercise.service;
 
 import java.util.List;
 
@@ -6,12 +6,12 @@ import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
 
-import com.hotspot.livfit.exercise.entity.entity.Lunge;
-import com.hotspot.livfit.exercise.entity.entity.Pushup;
-import com.hotspot.livfit.exercise.entity.entity.Squat;
-import com.hotspot.livfit.exercise.entity.repository.LungeRepository;
-import com.hotspot.livfit.exercise.entity.repository.PushupRepository;
-import com.hotspot.livfit.exercise.entity.repository.SquatRepository;
+import com.hotspot.livfit.exercise.entity.Lunge;
+import com.hotspot.livfit.exercise.entity.Pushup;
+import com.hotspot.livfit.exercise.entity.Squat;
+import com.hotspot.livfit.exercise.repository.LungeRepository;
+import com.hotspot.livfit.exercise.repository.PushupRepository;
+import com.hotspot.livfit.exercise.repository.SquatRepository;
 import com.hotspot.livfit.user.util.JwtUtil;
 
 @Service
